feat: implement multiplatform support
