+++ src/babel/types.ts
@@ -1,3 +1,4 @@
 export interface PluginOptions {
   icons?: string[];
+  entry?: string;
 }
