+++ src/babel/utils.ts
@@ -13,9 +13,16 @@ import IconNotFoundError from '../errors/icon-not-found.error';
 const iconsAsObject: Record<string, PartialExtendedIconifyIcon> = {};
 const collectionsAsObject: Record<string, IconifyJSON> = {};
 
-export const isIconifyFile = (plugin: b.PluginPass) =>
-  plugin.filename?.includes('Iconify') &&
-  plugin.file.code.includes('@@iconify-code-gen');
+export const isIconifyFile = (plugin: b.PluginPass) => {
+  const { entry } = plugin.opts as PluginOptions;
+
+  if (entry) return plugin.filename?.endsWith(entry);
+
+  return (
+    plugin.file.code.includes('@@iconify-code-gen') ||
+    plugin.filename?.includes('App.js')
+  );
+};
 
 export const loadIcon = (iconName: string) => {
   const iconDetails = stringToIcon(iconName);
@@ -39,8 +46,6 @@ export const loadIcon = (iconName: string) => {
   if (!icon) throw IconNotFoundError(iconName);
 
   iconsAsObject[iconName] = icon;
-
-  return icon;
 };
 
 export const loadIcons = (plugin: b.PluginPass) => {
@@ -51,7 +56,7 @@ export const loadIcons = (plugin: b.PluginPass) => {
   Array.from(new Set(icons)).forEach(loadIcon);
 
   const ast = b.template.ast(
-    `global.__ICONIFY__ = ${JSON.stringify(iconsAsObject)}`
+    `globalThis.__ICONIFY__ = ${JSON.stringify(iconsAsObject)};`
   ) as b.types.Statement;
 
   plugin.file.path.node.body.unshift(ast);
@@ -59,7 +64,7 @@ export const loadIcons = (plugin: b.PluginPass) => {
 
 export const setPluginInstalled = (plugin: b.PluginPass) => {
   const ast = b.template.ast(
-    `global.__ICONIFY_PLUGIN_LOADED__ = true;`
+    `globalThis.__ICONIFY_PLUGIN_LOADED__ = true;`
   ) as b.types.Statement;
 
   plugin.file.path.node.body.unshift(ast);
